feat(日志):一体机登录新增日志
一体机信息登录日志录入，提供一体机访问记录及次数查询

+++ src/main/java/com/zzj/mapperDM/DmUserAccountMapper.java
@@ -32,7 +32,7 @@ public interface DmUserAccountMapper {
     List<ExamResDTO> getExamList(Integer rowCount);
 
     List<ExamConfigDetailResDTO> getExamConf(Long userId);
-    Integer addRecord(ExamRecordReqDTO examRecordReqDTO);
+    Long addRecord(ExamRecordReqDTO examRecordReqDTO);
 
     Integer addRecordDetail(List<RecordDetailReqDTO> list,Integer reId);
